BACKEND/UserManagement/ViewUsers/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pymysql==1.0.3
python-dotenv==1.0.0
pydantic==1.10.7
requests==2.31.0
